Final removal of weaviate-specific logic
